backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.7
pytest==7.4.4
psycopg2-binary==2.9.11
pgvector==0.2.5
httpx>=0.27.0
trafilatura==1.7.0
readability-lxml==0.8.1
serpapi>=0.1.5
python-dotenv>=1.0.0
# Auth + user history
pymongo>=4.9.0
bcrypt>=4.2.0
PyJWT>=2.9.0
# Local embedding fallback (used when Gemini quota exhausted)
fastembed>=0.4.0
